Build 23020101

+++ ESP32/DIY-Flow-Bench/mafData/DELPHI_AF10118.cpp
@@ -71,7 +71,7 @@ int Maf::mafUnits() {
  * @brief mafLookupTable
  * @note Global vector of vectors containing MAF>Flow key>value pairs
  ***/
- std::vector<std::vector<int>> mafLookupTable = {{ 
+ std::vector<std::vector<long>> mafLookupTable = {{ 
 {0,0},
 {600,5620},
 {800,8982},
